dacn-dtgplx/PythonScripts/requirement.txt
# OCR
pytesseract==0.3.10
Pillow==10.2.0

# xử lý vector, embedding
numpy==1.26.4
scikit-learn==1.4.2
openai>=1.51.0

# HuggingFace transformer (embedding model)
transformers==4.39.3
tokenizers==0.15.2

# Sentence transformer (nhúng câu)
sentence-transformers==2.5.1

# Tiện ích
tqdm==4.66.2
